+++ plugins/amazonq/chat/jetbrains-community/src/software/aws/toolkits/jetbrains/services/cwc/editor/context/project/ProjectContextEditorListener.kt
@@ -5,9 +5,7 @@ package software.aws.toolkits.jetbrains.services.cwc.editor.context.project
 import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.fileEditor.FileEditorManagerListener
 import com.intellij.openapi.vfs.VirtualFile
-import software.aws.toolkits.jetbrains.services.amazonq.apps.AmazonQAppInitContext
 import software.aws.toolkits.jetbrains.services.codewhisperer.settings.CodeWhispererSettings
-import software.aws.toolkits.jetbrains.services.cwc.inline.ChatCaretListener
 
 class ProjectContextEditorListener() : FileEditorManagerListener {
     override fun fileClosed(source: FileEditorManager, file: VirtualFile) {
