+++ plugins/amazonq/chat/jetbrains-community/src/software/aws/toolkits/jetbrains/services/cwc/inline/listeners/InlineChatSelectionListener.kt
@@ -1,10 +1,11 @@
 // Copyright 2024 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.services.cwc.inline
+package software.aws.toolkits.jetbrains.services.cwc.inline.listeners
 
 import com.intellij.openapi.editor.event.SelectionEvent
 import com.intellij.openapi.editor.event.SelectionListener
+import software.aws.toolkits.jetbrains.services.cwc.inline.InlineChatEditorHint
 import java.awt.Point
 
 class InlineChatSelectionListener : SelectionListener {
